Re-tokenise combined SQL statements prior to parsing, to insulate ourselves from faulty assumptions and from character-counting errors in the Percona parser
The actual problem reported in this case was that parsing failed if a field was updated with a value consisting of a single comma, as the parse_csv function naïvely splits on commas and so thinks that the correct result is two undefined values - which then confused the calling parse_set function.  The underlying issue is still present, but should hopefully now never be triggered as unexpected values should no longer ever be seen by the parser.